code structure swaps create_pool
