+++ zrml/swaps/fuzz/pool_exit_with_exact_pool_amount.rs
@@ -3,16 +3,11 @@
 use libfuzzer_sys::fuzz_target;
 use zrml_swaps::mock::{ExtBuilder, Origin, Swaps};
 
-mod data_structs;
-use data_structs::ExactAmountData;
-mod helper_functions;
-use helper_functions::{construct_asset, _CREATE_POOL_FAILURE};
+mod utils;
 use orml_traits::MultiCurrency;
-use zeitgeist_primitives::{
-    constants::MinLiquidity,
-    traits::Swaps as SwapsTrait,
-    types::{Asset, ScoringRule, SerdeWrapper},
-};
+use utils::{construct_asset, ExactAmountData};
+use zeitgeist_primitives::constants::MinLiquidity;
+// TODO use zeitgeist_primitives::types::{Asset, SerdeWrapper};
 use zrml_swaps::mock::Shares;
 
 fuzz_target!(|data: ExactAmountData| {
@@ -27,32 +22,24 @@ fuzz_target!(|data: ExactAmountData| {
                 MinLiquidity::get(),
             );
         }
+        /*
+        TODO
+
         let _ = Shares::deposit(
             Asset::PoolShare(SerdeWrapper(0)),
             &data.pool_creation.origin,
             data.pool_amount,
         );
+        */
 
-        match Swaps::create_pool(
-            data.pool_creation.origin,
-            data.pool_creation.assets.into_iter().map(construct_asset).collect(),
-            construct_asset(data.pool_creation.base_asset),
-            data.pool_creation.market_id,
-            ScoringRule::CPMM,
-            Some(data.pool_creation.swap_fee),
-            Some(data.pool_creation.weights),
-        ) {
-            Ok(pool_id) => {
-                let _ = Swaps::pool_exit_with_exact_pool_amount(
-                    Origin::signed(data.origin),
-                    pool_id,
-                    construct_asset(data.asset),
-                    data.pool_amount,
-                    data.asset_amount,
-                );
-            }
-            Err(e) => panic!("{_CREATE_POOL_FAILURE} {:?}", e),
-        }
+        let pool_id = data.pool_creation._create_pool();
+        let _ = Swaps::pool_exit_with_exact_pool_amount(
+            Origin::signed(data.origin),
+            pool_id,
+            construct_asset(data.asset),
+            data.pool_amount,
+            data.asset_amount,
+        );
     });
     let _ = ext.commit_all();
 });
