+++ zrml/swaps/fuzz/pool_join.rs
@@ -2,12 +2,10 @@
 
 use libfuzzer_sys::fuzz_target;
 
+use utils::GeneralPoolData;
 use zrml_swaps::mock::{ExtBuilder, Origin, Swaps};
-mod data_structs;
-use data_structs::GeneralPoolData;
-use zeitgeist_primitives::{traits::Swaps as SwapsTrait, types::ScoringRule};
-mod helper_functions;
-use helper_functions::{construct_asset, _CREATE_POOL_FAILURE};
+mod utils;
+use utils::construct_asset;
 use zeitgeist_primitives::constants::MinLiquidity;
 use zrml_swaps::mock::Shares;
 
@@ -25,27 +23,10 @@ fuzz_target!(|data: GeneralPoolData| {
                 MinLiquidity::get(),
             );
         }
-
-        match Swaps::create_pool(
-            data.pool_creation.origin,
-            data.pool_creation.assets.into_iter().map(construct_asset).collect(),
-            construct_asset(data.pool_creation.base_asset),
-            data.pool_creation.market_id,
-            ScoringRule::CPMM,
-            Some(data.pool_creation.swap_fee),
-            Some(data.pool_creation.weights),
-        ) {
-            Ok(pool_id) => {
-                // join a pool with a valid pool id
-                let _ = Swaps::pool_join(
-                    Origin::signed(data.origin),
-                    pool_id,
-                    data.pool_amount,
-                    data.assets,
-                );
-            }
-            Err(e) => panic!("{_CREATE_POOL_FAILURE} {:?}", e),
-        }
+        let pool_id = data.pool_creation._create_pool();
+        // join a pool with a valid pool id
+        let _ =
+            Swaps::pool_join(Origin::signed(data.origin), pool_id, data.pool_amount, data.assets);
     });
 
     let _ = ext.commit_all();
