+++ zrml/swaps/fuzz/swap_exact_amount_out.rs
@@ -3,14 +3,10 @@
 use libfuzzer_sys::fuzz_target;
 use zrml_swaps::mock::{ExtBuilder, Origin, Swaps};
 
-mod data_structs;
-use data_structs::SwapExactAmountData;
-mod helper_functions;
-use helper_functions::{construct_asset, _CREATE_POOL_FAILURE};
+mod utils;
 use orml_traits::MultiCurrency;
-use zeitgeist_primitives::{
-    constants::MinLiquidity, traits::Swaps as SwapsTrait, types::ScoringRule,
-};
+use utils::{construct_asset, SwapExactAmountData};
+use zeitgeist_primitives::constants::MinLiquidity;
 use zrml_swaps::mock::Shares;
 
 fuzz_target!(|data: SwapExactAmountData| {
@@ -25,34 +21,17 @@ fuzz_target!(|data: SwapExactAmountData| {
                 MinLiquidity::get(),
             );
         }
-
-        match Swaps::create_pool(
-            data.pool_creation.origin,
-            data.pool_creation.assets.into_iter().map(construct_asset).collect(),
-            construct_asset(data.pool_creation.base_asset),
-            data.pool_creation.market_id,
-            ScoringRule::CPMM,
-            Some(data.pool_creation.swap_fee),
-            Some(data.pool_creation.weights),
-        ) {
-            Ok(pool_id) => {
-                let _ = Shares::deposit(
-                    construct_asset(data.asset_in),
-                    &data.origin,
-                    data.asset_amount_in,
-                );
-                let _ = Swaps::swap_exact_amount_out(
-                    Origin::signed(data.origin),
-                    pool_id,
-                    construct_asset(data.asset_in),
-                    data.asset_amount_in,
-                    construct_asset(data.asset_out),
-                    data.asset_amount_out,
-                    data.max_price,
-                );
-            }
-            Err(e) => panic!("{_CREATE_POOL_FAILURE} {:?}", e),
-        }
+        let pool_id = data.pool_creation._create_pool();
+        let _ = Shares::deposit(construct_asset(data.asset_in), &data.origin, data.asset_amount_in);
+        let _ = Swaps::swap_exact_amount_out(
+            Origin::signed(data.origin),
+            pool_id,
+            construct_asset(data.asset_in),
+            data.asset_amount_in,
+            construct_asset(data.asset_out),
+            data.asset_amount_out,
+            data.max_price,
+        );
     });
     let _ = ext.commit_all();
 });
